requirements.txt
asgiref==3.11.0
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
crispy-bootstrap5==2025.6
cryptography==46.0.3
cycler==0.12.1
diff-match-patch==20241021
Django==5.2.8
django-appconf==1.2.0
django-axes==8.0.0
django-cors-headers==4.9.0
django-countries==8.1.1
django-crispy-forms==2.5
django-extensions==4.1
django-filter==25.2
django-import-export==4.3.14
django-jazzmin==3.0.1
django-phonenumber-field==8.3.0
django-phonenumbers==1.0.1
django-rest-framework==0.1.0
django-restframework==0.0.1
django-select2==8.4.3
django-simple-history==3.10.1
django-tabbed-admin==1.0.3
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Faker==38.2.0
fonttools==4.60.1
gunicorn==23.0.0
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.5
packaging==25.0
pandas==2.3.3
phonenumbers==9.0.19
pillow==12.0.0
psycopg==3.2.13
psycopg-binary==3.2.13
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
reportlab==4.4.5
six==1.17.0
sqlparse==0.5.3
tablib==3.9.0
typing_extensions==4.15.0
tzdata==2025.2
whitenoise==6.11.0
